Divide redux architecture
