+++ src/components/counter.js
@@ -1,18 +1,24 @@
-// src/components/counter.js
-import React, { Component } from 'react';
+// container
+// initial
+import React from "react";
+import "../App.css";
+import PropTypes from "prop-types";
 
-class Counter extends Component {
-  render() {
-   let { counter, dispatch_increment, dispatch_decrement } = this.props;
-   console.log(counter.note)
-   return (
-     <div>
-       <div>{counter.count}</div>
-       <button onClick={dispatch_increment}>+</button>
-       <button onClick={dispatch_decrement}>-</button>
-     </div>
-   );
-  }
+function Counter({ counter, dispatch_increment, dispatch_decrement }) {
+  return (
+    <div>
+      <div>{counter.count}</div>
+      <button onClick={dispatch_increment}>+</button>
+      <button onClick={dispatch_decrement}>-</button>
+    </div>
+  );
 }
+
+// proptypes
+Counter.propTypes = {
+  counter: PropTypes.object.isRequired,
+  dispatch_increment: PropTypes.func.isRequired,
+  dispatch_decrement: PropTypes.func.isRequired,
+};
 
 export default Counter;
